requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.32
data==0.4
decorator==5.1.1
docker-pycreds==0.4.0
fasteners==0.17.3
fonttools==4.32.0
funcsigs==1.0.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.4
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.23.0
-e git+https://github.com/maximecb/gym-miniworld@b98ba020f1d802df75e4c73b0d0ab656ff917453#egg=gym_miniworld
gym-notices==0.0.5
idna==3.3
imageio==2.20.0
importlib-metadata==4.11.2
kiwisolver==1.4.2
latex==0.7.0
Markdown==3.3.6
matplotlib==3.5.1
mujoco==2.2.1
mujoco-py==2.1.2.14
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyglet==1.5.22
PyOpenGL==3.1.6
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
sentry-sdk==1.5.7
setproctitle==1.2.2
shortuuid==1.0.8
shutilwhich==1.1.0
six==1.16.0
smmap==5.0.0
tempdir==0.7.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
torch==1.10.2
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
wandb==0.12.11
Werkzeug==2.0.3
yaspin==2.1.0
zipp==3.7.0
